+++ src/cynn/__init__.py
@@ -1,10 +1,12 @@
-from ._core import TinnNetwork, GenannNetwork, FannNetwork, FannNetworkDouble, square, seed
+from ._core import TinnNetwork, GenannNetwork, FannNetwork, FannNetworkDouble, CNNNetwork, CNNLayer, square, seed
 
 __all__ = [
     'TinnNetwork',
     'GenannNetwork',
     'FannNetwork',
     'FannNetworkDouble',
+    'CNNNetwork',
+    'CNNLayer',
     'square',
     'seed',
 ]